backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["pibragimov <pibragimov@whoosh.bike>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10,<3.13"
sqlalchemy = { extras = ["asyncio"], version = "^2.0.21" }
psycopg2-binary = "^2.9.3"
ujson = "^5.8.0"
fastapi = "^0.103.2"
pydantic-settings = "^2.0.3"
uvicorn = "^0.23.2"
pyhumps = "^3.8.0"
loguru = "^0.7.2"
asyncpg = "^0.28.0"
pytest = "^7.4.2"
pytest-asyncio = "^0.21.1"
fastapi-cache2 = { extras = ["redis"], version = "^0.2.1" }
jupyterlab = "^4.0.7"
numpy = "^1.26.0"
pandas = "^2.1.1"
ulid-py = "^1.1.0"
geoalchemy2 = "^0.14.1"
tqdm = "^4.66.1"
shapely = "^2.0.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
